viikko03-Viikko03_063.MerkkijononKaantaminen/src/MerkkijononKaantaminen.java

import java.util.Scanner;

public class MerkkijononKaantaminen {

    public static String kaanna(String merkkijono) {
        // Kirjoita koodia tähän.
        // Metodin on siis palautettava käännetty merkkijono returnilla!
        
        String palautus = "";
        
        for (int i = merkkijono.length()-1; i >= 0; i--) {
            palautus += merkkijono.charAt(i);
        }
        return palautus;
    }

    public static void main(String[] args) {
        Scanner lukija = new Scanner(System.in);
        System.out.print("Anna merkkijono: ");
        String merkkijono = lukija.nextLine();
        System.out.println("Väärinpäin: " + kaanna(merkkijono));
    }
}
